next snapshot

+++ scripts/issue-service-offering-credential.sh
@@ -1,7 +1,12 @@
 #!/bin/sh
+
+if [ "$#" -ne 3 ]; then echo "Please provide a participant name, a service offering name and ID URL"
+    exit -1
+fi
+
 docker run -it -v "$(pwd)/../dataset:/ext" --rm onboardingcli/onboardingcli vc create \
---claims-file /ext/claims/invoice-service-offering.json --template-file /ext/templates/template.json \
---trusted-issuer-file /ext/identities/acme/acme.json --vc-version 2 \
---subject-did "https://service-offerings.example.org/commercial-invoice-1" \
---issuers-dir /ext/identities --templates-dir /ext/templates --vc-dir /ext/credentials/evidences \
+--claims-file "/ext/claims/service-offerings/$2-service-offering.json" --template-file /ext/templates/template.json \
+--trusted-issuer-file "/ext/identities/$1/$1.json" --vc-version 2 \
+--subject-did "$3" \
+--issuers-dir /ext/identities --templates-dir /ext/templates --vc-dir /ext/credentials/evidences/service-offerings \
 --valid-for 20000h
